+++ public/documentation/js/Settings.js
@@ -21,7 +21,7 @@ export const Settings = {
     DOCBASE: 'https://raw.githubusercontent.com/catenax-ng/tx-portal-assets',
     SRCBASE: 'https://github.com/catenax-ng/tx-portal-assets',
     DEFAULT_BRANCH: 'main',
-    ROOT: 'docs',
+    DEFAULT_ROOT: 'docs',
 }
 
 export const Patterns = {